requirements.txt
flask==2.2.3
pymongo==4.3.3
python-dotenv==1.0.0
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
werkzeug==2.2.3
Flask-PyMongo==2.3.0
bcrypt==4.0.1
python-dateutil==2.8.2